V5.9.187c: Fix two compile errors
1. Executor.kt:804 — it.score is Double, coerceIn(0,100) has Int literals
   Fix: coerceIn(0.0, 100.0).toInt()

2. V3EngineManager.kt:356 — avgLossPct does not exist in LifetimeSnapshot
   LifetimeSnapshot fields: totalSells/Wins/Losses/Scratches, winRate, avgWinPct, realizedPnlSol
   Fix: payoffRatio = (avgWinPct / 8.0).coerceIn(0.5, 5.0)
   (8% = typical stop-loss. If bot wins avg 12%, payoff ratio = 1.5. Sensible proxy.)

+++ lifecycle_apk/app/src/main/kotlin/com/lifecyclebot/engine/Executor.kt
@@ -801,7 +801,7 @@ class Executor(
                     
                     val mode = try { ModeRouter.classify(ts).tradeType.name } catch (_: Exception) { "UNKNOWN" }
                     val score = ts.trades.lastOrNull { it.side == "BUY" }?.let { 
-                        it.score.coerceIn(0, 100)  // V5.9.186: was price*100 = always 100 (wrong) 
+                        it.score.coerceIn(0.0, 100.0).toInt()  // V5.9.186: was price*100 = always 100 (wrong) 
                     } ?: 50
                     val confidence = (trade.pnlPct ?: 0.0).toInt().coerceIn(0, 100)
